+++ tests/integration/api/mod.rs
@@ -2,4 +2,5 @@ mod account;
 mod auth;
 mod clients;
 mod mail;
+mod probes;
 mod two_factor;